Fixed load Issues, Clean up
fixed the issues due to loading and not updating all respective values. Now able to load a saved file and continue with no issues.

Cleaned up some of the code. Removed BoxManager subclass as it was no longer reasonable to have. Refactored names to be more consistent and logical for their resoective tasks.