Encryption and Save on Close
Added file encryption and cleaner way of appending the desired extension onto the file. Files are now encrypted on save, and decrypted then immediately re-encrypted on load. Added a flag that checks if any cells or edges were added or removed to mark the file as changed. This is used to decide to promt the user on exit to save changes or not.